Add ability to expand/collapse folders and show contents
I had to spend a bit of time figuring out how folder paths worked.

The key insight is that if the has constructor of Subfolder then that
is an instruction to descend into that subfolder. Otherwise it is an
End and we need to stop at that node. It is always relative to the
node you're currently looking at.